Move session.ts to services/session.server.ts
